requirements.txt
scrapy
requests
beautifulsoup4
newspaper3k
anthropic
lxml_html_clean
python-dotenv
elevenlabs
pygame
flask
gunicorn